Pacify PVS-Studio by changing the type rather than doing a cast, a tweak suggested by David Blaikie.
llvm-svn: 162286

+++ llvm/include/llvm/Bitcode/BitstreamWriter.h
@@ -172,7 +172,7 @@ public:
     if ((uint32_t)Val == Val)
       return EmitVBR((uint32_t)Val, NumBits);
 
-    uint64_t Threshold = uint64_t(1U << (NumBits-1));
+    uint32_t Threshold = 1U << (NumBits-1);
 
     // Emit the bits with VBR encoding, NumBits-1 bits at a time.
     while (Val >= Threshold) {
